+++ app/Dockerfile
@@ -1,4 +1,4 @@
-FROM python:3.12-slim
+FROM python:3.13-slim
 WORKDIR /app
 COPY pyproject.toml poetry.lock /app/
 RUN pip install poetry
